Drop "should" clauses
Don't say "should", (e.g. "A notebook should begin...") just drop the "should" clause and make it a declarative statement: "Begin the notebook with...". Be prescriptive ("Provide a brief description of each file.") rather than descriptive ("A description of what a file is would be helpful for a new user.").